Use json5 filename to match format
Only json5 format allows comments in json files. Renovate is about to
remove the support for json5 parsing in .json files, instead we need to
explicitly use .json5 filenames to use that format.

```
WARN: File contents are invalid JSON but parse using JSON5.
Support for this will be removed in a future release so please
change to a support .json5 file name or ensure correct JSON syntax.
```